Rename and refactor extenion methods to be easier to read
